Re-add flask-flash widget to signin page
The flask-flash widget was originally part of the `navbar` block,
which meant that back in e2caa31 as part of #39, hiding the navbar
block on the signin page disabled the flash widget. This led to a
garbage collection error (probably due to poor programming in the
flask-flash package).

This has been corrected by giving flask-flash it's own block
outside of the navbar block.